+++ Main_Core/libft/ft_calloc.c
@@ -6,20 +6,19 @@
 /*   By: nkarpeko <nkarpeko@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/14 15:20:11 by nkarpeko          #+#    #+#             */
-/*   Updated: 2023/04/14 15:22:11 by nkarpeko         ###   ########.fr       */
+/*   Updated: 2023/04/14 15:52:37 by nkarpeko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_calloc(size_t num, size_t size)
+void	*ft_calloc(int num, int size)
 {
 	void	*ptr;
 
 	ptr = malloc(num * size);
 	if (ptr != NULL)
 	{
-		// Zero out the allocated memory
 		ft_memset(ptr, 0, num * size);
 	}
 	return (ptr);
